COMMON: Restore old GWIM behavior so that it never matches ME.
The standard behavior of "THROW object AT ME" is to kill the player. The
syntax for THROW tells it to pick an actor as the indirect object, if
none is specified. The ME object is an actor, and in Zork I and III it's
often the only actor present. (In Zork II, you also have the
GLOBAL-PRINCESS object.)

That meant that "THROW object" was often a lethal command, which seems
unfair even for Zork. In older versions, GWIM had a special case that
prevented it from matching the ME object, but that was removed in the
more recent version of the parser. This simply restores the old
behavior.