COMMON: Restore old V-OPEN / V-CLOSE behavior
Objects need CONTBIT to be openable/closeable by default. This behavior
goes back all the way to Mainframe Zork, and is necessary for Zork II
and III where certain objects are "doors" but not meant to be
opened/closed. The CLEFT in Zork III is one such example.

While it's probably not needed for Zork I, it don't think it will do any
harm there.