ZORK1: Fix printing of objects in the room below when you're up a tree
The old behavior was to check if there were two objects, not one, in the
room below. This apparently goes all the way back to Mainframe Zork,
where the tree itself appears to have been a proper object in the room
below. In Zork I, the tree is a fake object.