ZORK3: Don't allow the figure to block your way when he's wounded
The game goes to the trouble of not setting BLOCKED-DIR when he's
weakened, and describing him as "probably not capable of hindering your
movement". But once he was blocking a direction, he would do so even
when wounded. We could clear BLOCKED-DIR, but let's not do that. He
could recover during the fight.