ZORK2: Add SMOKEY as an adjective for TUNNEL
Zork II explicitly refers to a tunnel as "smokey" rather than "smoky".
Zork III also has this object, but since it never refers to anything as
"smokey" except for "Smokey the Bear", it can stay as it is. Zork I
doesn't have a global TUNNEL object, perhaps becaues a sizable chunk of
it takes place above ground.