ZORK1: Tweaked the slide a bit
You can no longer put objects in the slide from the cellar, since the
message for it makes no sense. Instead, print the message that would
have been printed if you tried to put a non-takeable object, which I'm
not sure is even possible to do. In Mainframe Zork, the slide was not
accessible from the cellar anyway, so this is fairly consistent with the
old behavior.

Added a blank line between "You tumble down the slide...." and the room
description, because that's how such messages are usually printed.