ZORK1: I-MAINT-ROOM tweaks
Don't print messages about how high the water is if you're in the boat.
It's enough with the message telling you that the boat is carried away
and over the falls.

Don't drown until the water is "high in your lungs". This is how
Mainframe Zork worked, and I don't know why it was changed. The texts
were still there, after all.