ZORK1: Print room description when entering Dome Room as a ghost
When you are blown into the Torch Room as a ghost, the game used to
print the room description. The code moves the player manually, rather
than using GOTO. I'm guessing GOTO would have printed the room
description twice, but these days GOTO suppresses the room description
if M-ENTER changes the player location.

So it should be safe to use GOTO here, and doing so restores the old
behavior.