ZORK1: Print correct message when jumping in the Dome Room
The old behavior was still to kill you, but the custom message for it
was in M-ENTER, not M-BEG, so it was never shown. And since enabling the
custom code would kill you when jumping over anything, I've narrowed it
to jumping without any specified destination, and jumping over the
railing.

This is somewhat based on CANYON-VIEW-F.