ZORK1: When waving the sceptre, check if the pot is still invisible
Waving the sceptre a second time doesn't make the pot go away, so it
does not seem right that the game tells you that the pot appears when
you wave the sceptre a third time. Also, only turn the pot visible if
it's still at the end of the rainbow. Not, say, when the thief is
carrying it.

Of course, we could also make the pot go away if it hasn't been touched.
But that's a more invasive change.