ZORK1: Fix some inconsistencies in SAND-FUNCTION
Before, you would see the scarab if it was invisible, even if it wasn't
there any more (e.g. the thief has it). Also, digging in the sand when
the scarab wasn't invisible (i.e. after a resurrection) would print the
default message on the turn the scarab would have been revealed. Now it
collapses the hole one move earlier instead.