ZORK1: Fix strange behavior when burning the leaves
The old behavior was that it was only safe to burn the leaves when they
were in the room with you. They never were, since the leaves were
removed before that check was made. But even after fixing that, you
would kill yourself if you put the leaves in a container in the room
(e.g. the table or the coffin) and burned them there. I've changed it to
a HELD? check.

Finally, burning the leaves to reveal the grating said that the leaves
were "moved". I've changed it to say "gone".