ZORK1: Don't clear TOUCHBIT on maze rooms while walking
Now, I do understand that this was probably done so that you couldn't
tell whether you had already visited that part of the maze when using
brief mode. But it also meant that the thief almost never picked up
stuff in the maze, and the ROB-MAZE routine was probably never called.

Of course this means that the maze is now more annoying to map. But the
"My, I wonder what this fine <object> is doing here." message when the
thief picks up junk in the maze has been a feature since forever. I
think it was mentioned in the very first article I ever read about Zork.
It'd be a shame to lose it.

This seems to match the behavior in the Solid Gold version, so I guess
someone at Infocom felt the same way.