ZORK1: Remove default message when talking to the bat
The first version of Zork I where you could talk to the bat (or at least
the first one where I could figure out how to) did not have this default
message, so it was a regression caused by the action routine returning
false. Also fixed the extra blank line that would appear at the end.