ZORK1: Several fixes to GRATE-FUNCTION
I'm sorry, but this one had a lot of rough edges.

Locking and unlocking the grate did not check if the grate was already
locked or unlocked, you could lock it even if it was closed, and you
could lock it without the key. I've rewritten that to be more robust.

There were a few places where it incorrectly checked if you were in
CLEARING, rather than GRATING-CLEARING. I think the only visible effect
of that was that if you opened the grating from above, you were told
there were trees above. I've fixed that.

Closing the grating from below now informs the player if the room went
dark.

You could put objects through the grating from below, and it still said
that the object fell into the darkness below. I've changed that so that
you can only do it from above. I've also removed the size check of the
object when the grating is open. After all, the pile of leaves can fall
through on its own when you open the grating, so there's no reason you
shouldn't be able to put it through yourself.